# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  alioth-hrc_internal_model.GiftDirection:
    enum:
    - give
    - receive
    type: string
    x-enum-varnames:
    - DirectionGive
    - DirectionReceive
  alioth-hrc_internal_model.GiftRecord:
    properties:
      amount_cents:
        type: integer
      category_id:
        type: integer
      contact_id:
        type: integer
      created_at:
        type: string
      direction:
        $ref: '#/definitions/alioth-hrc_internal_model.GiftDirection'
      id:
        type: integer
      ledger_id:
        type: integer
      note:
        type: string
      occurred_on:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  alioth-hrc_internal_service.GiftRecordCreate:
    properties:
      amount_cents:
        minimum: 1
        type: integer
      category_id:
        type: integer
      contact_id:
        type: integer
      direction:
        enum:
        - give
        - receive
        type: string
      ledger_id:
        type: integer
      note:
        maxLength: 1000
        type: string
      occurred_on:
        description: YYYY-MM-DD
        type: string
    required:
    - amount_cents
    - category_id
    - contact_id
    - direction
    - ledger_id
    - occurred_on
    type: object
  alioth-hrc_internal_service.GiftRecordUpdate:
    properties:
      amount_cents:
        minimum: 1
        type: integer
      category_id:
        type: integer
      contact_id:
        type: integer
      direction:
        enum:
        - give
        - receive
        type: string
      ledger_id:
        type: integer
      note:
        maxLength: 1000
        type: string
      occurred_on:
        type: string
    required:
    - amount_cents
    - category_id
    - contact_id
    - direction
    - ledger_id
    - occurred_on
    type: object
...
  internal_handler.ErrorJSON:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.GiftRecordListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/alioth-hrc_internal_model.GiftRecord'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handler.OKStatus:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: 人情往来记账后端：认证、联系人、账本、礼金分类与流水、统计等。
  title: alioth-hrc API
  version: "1.0"
paths:
...
  /gift-records:
    get:
      parameters:
      - description: 联系人 ID
        in: query
        name: contact_id
        type: integer
      - description: 账本 ID
        in: query
        name: ledger_id
        type: integer
      - description: 分类 ID
        in: query
        name: category_id
        type: integer
      - description: '方向: give 或 receive'
        in: query
        name: direction
        type: string
      - description: 开始日期 YYYY-MM-DD
        in: query
        name: from_date
        type: string
      - description: 结束日期 YYYY-MM-DD
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.GiftRecordListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 礼金流水列表
      tags:
      - gift-records
    post:
      consumes:
      - application/json
      parameters:
      - description: 流水
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/alioth-hrc_internal_service.GiftRecordCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.GiftRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 创建礼金流水
      tags:
      - gift-records
  /gift-records/{id}:
    delete:
      parameters:
      - description: 流水 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.OKStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 删除礼金流水
      tags:
      - gift-records
    get:
      parameters:
      - description: 流水 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.GiftRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 礼金流水详情
      tags:
      - gift-records
    put:
      consumes:
      - application/json
      parameters:
      - description: 流水 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 流水
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/alioth-hrc_internal_service.GiftRecordUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.GiftRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 更新礼金流水
      tags:
      - gift-records
  /gift-records/export.csv:
    get:
      parameters:
      - description: 联系人 ID
        in: query
        name: contact_id
        type: integer
      - description: 账本 ID
        in: query
        name: ledger_id
        type: integer
      - description: 分类 ID
        in: query
        name: category_id
        type: integer
      - description: 方向
        in: query
        name: direction
        type: string
      - description: 开始日期 YYYY-MM-DD
        in: query
        name: from_date
        type: string
      - description: 结束日期 YYYY-MM-DD
        in: query
        name: to_date
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV 文件
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 导出礼金流水 CSV
      tags:
      - gift-records
...
securityDefinitions:
  BearerAuth:
    description: 访问令牌，格式：Bearer {access_token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"